Create new default branch

+++ Biochemical Lab/Assets/Photon/PhotonVoice/PhotonVoiceApi/Platforms/Unity/UnityAudioOut.cs
@@ -1,7 +1,4 @@
 ﻿using UnityEngine;
-using System.Collections.Generic;
-using System;
-using System.Runtime.InteropServices;
 
 namespace Photon.Voice.Unity
 {
@@ -18,7 +15,7 @@ namespace Photon.Voice.Unity
             this.source = audioSource;
         }
 
-        override public int OutPos { get { return source.timeSamples; } }
+        override public long OutPos { get { return source.timeSamples; } }
 
         override public void OutCreate(int frequency, int channels, int bufferSamples)
         {
